+++ src/array/array_metadata_options.rs
@@ -1,11 +1,23 @@
-use crate::config::{MetadataOptionsEraseVersion, MetadataOptionsStoreVersion};
+use crate::config::{global_config, MetadataOptionsEraseVersion, MetadataOptionsStoreVersion};
 
 /// Options for writing array metadata.
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone)]
 pub struct ArrayMetadataOptions {
     experimental_codec_store_metadata_if_encode_only: bool,
     store_version: MetadataOptionsStoreVersion,
     erase_version: MetadataOptionsEraseVersion,
+    include_zarrs_metadata: bool,
+}
+
+impl Default for ArrayMetadataOptions {
+    fn default() -> Self {
+        Self {
+            experimental_codec_store_metadata_if_encode_only: false,
+            store_version: MetadataOptionsStoreVersion::default(),
+            erase_version: MetadataOptionsEraseVersion::default(),
+            include_zarrs_metadata: global_config().include_zarrs_metadata(),
+        }
+    }
 }
 
 impl ArrayMetadataOptions {
@@ -51,6 +63,18 @@ impl ArrayMetadataOptions {
         erase_version: MetadataOptionsEraseVersion,
     ) -> &mut Self {
         self.erase_version = erase_version;
+        self
+    }
+
+    /// Get the [include zarrs metadata](crate::config::Config#include-zarrs-metadata) configuration.
+    #[must_use]
+    pub fn include_zarrs_metadata(&self) -> bool {
+        self.include_zarrs_metadata
+    }
+
+    /// Set the [include zarrs metadata](crate::config::Config#include-zarrs-metadata) configuration.
+    pub fn set_include_zarrs_metadata(&mut self, include_zarrs_metadata: bool) -> &mut Self {
+        self.include_zarrs_metadata = include_zarrs_metadata;
         self
     }
 }